refactore Controllers and some connected Services; Create ControllerHelpers;

+++ wizards/Wizard.Services/HeroService/HeroService.cs
@@ -4,6 +4,7 @@ using Wizards.Core.Model;
 using Wizards.Services.Factories;
 using Wizards.Services.PlayerService;
 using Wizards.Services.Validation;
+using Wizards.Services.Validation.Elements;
 
 namespace Wizards.Services.HeroService;
 
@@ -22,7 +23,6 @@ public class HeroService : IHeroService
         _propertiesFactory = propertiesFactory;
     }
 
-
     public async Task Add(int playerId, Hero hero)
     {
         await _heroValidator.Validate(hero);
@@ -36,29 +36,29 @@ public class HeroService : IHeroService
         await _heroRepository.Add(player, hero);
     }
 
-    public async Task Delete(int id)
+    public async Task Delete(int id, string confirmNickName)
     {
         var hero = await Get(id);
+
+        if (hero.NickName != confirmNickName)
+        {
+            var message = new Dictionary<string, string>();
+            message.Add("ConfirmNickName", "Invalid Nick Name!");
+            throw new InvalidModelException(message);
+        }
+
         await _heroRepository.Remove(hero);
     }
 
     public async Task Update(int id, Hero hero)
     {
+        var heroToUpdate = await Get(id);
+        hero.Profession = heroToUpdate.Profession;
+        
         await _heroValidator.Validate(hero);
         
-        var heroToUpdate = await Get(id);
-
-        if (heroToUpdate.NickName != hero.NickName && await CanChangeNickName(heroToUpdate.Id))
-        {
-            heroToUpdate.NickName = hero.NickName;
-            heroToUpdate.Gold -= GetChangeNickNameCost();
-        }
-
-        if (heroToUpdate.AvatarImageNumber != hero.AvatarImageNumber && await CanChangeAvatar(heroToUpdate.Id))
-        {
-            heroToUpdate.AvatarImageNumber = hero.AvatarImageNumber;
-            heroToUpdate.Gold -= GetChangeAvatarCost();
-        }
+        ChangeNickName(heroToUpdate, hero.NickName);
+        ChangeAvatar(heroToUpdate, hero.AvatarImageNumber);
 
         await _heroRepository.Update(heroToUpdate);
     }
@@ -75,28 +75,6 @@ public class HeroService : IHeroService
         throw new NullReferenceException($"There is no Hero with id: {id}!");
     }
 
-    public async Task<bool> HasPlayerHero(ClaimsPrincipal user, int heroId)
-    {
-        var playerId = _playerService.GetId(user);
-        var player = await _playerService.Get(playerId);
-
-        var result = player.Heroes.Any(h => h.Id == heroId);
-        
-        return result;
-    }
-
-    public async Task<bool> CanChangeNickName(int id)
-    {
-        var hero = await Get(id);
-        return (hero.Gold >= GetChangeNickNameCost());
-    }
-
-    public async Task<bool> CanChangeAvatar(int id)
-    {
-        var hero = await Get(id);
-        return (hero.Gold >= GetChangeAvatarCost());
-    }
-
     public int GetChangeNickNameCost()
     {
         return 2500;
@@ -105,5 +83,51 @@ public class HeroService : IHeroService
     public int GetChangeAvatarCost()
     {
         return 2500;
+    }
+
+    private void ChangeNickName(Hero heroToUpdate, string nickName)
+    {
+        if (heroToUpdate.NickName == nickName)
+        {
+            return;
+        }
+
+        if (!CanChangeNickName(heroToUpdate))
+        {
+            var message = new Dictionary<string, string>();
+            message.Add("NickName", "You don't have enough gold!");
+            throw new InvalidModelException(message);
+        }
+
+        heroToUpdate.NickName = nickName;
+        heroToUpdate.Gold -= GetChangeNickNameCost();
+    }
+
+    private void ChangeAvatar(Hero heroToUpdate, int avatarNumber)
+    {
+        if (heroToUpdate.AvatarImageNumber == avatarNumber)
+        {
+            return;
+        }
+
+        if (!CanChangeAvatar(heroToUpdate))
+        {
+            var message = new Dictionary<string, string>();
+            message.Add("NickName", "You don't have enough gold!");
+            throw new InvalidModelException(message);
+        }
+
+        heroToUpdate.AvatarImageNumber = avatarNumber;
+        heroToUpdate.Gold -= GetChangeNickNameCost();
+    }
+
+    private bool CanChangeNickName(Hero hero)
+    {
+        return (hero.Gold >= GetChangeNickNameCost());
+    }
+
+    private bool CanChangeAvatar(Hero hero)
+    {
+        return (hero.Gold >= GetChangeAvatarCost());
     }
 }
